refactor: merge user.server.js to session.server.js
