+++ app/utils/session.server.js
@@ -2,22 +2,39 @@ import crypto from 'crypto'
 import bcrypt from 'bcryptjs'
 import { createCookieSessionStorage, redirect } from '@remix-run/node'
 import {
-  countUser,
-  getUser,
-  getUserWithProfile,
-  saveUser,
-  updateUserLogin
-} from '~/utils/user.server'
-import { mongoConnect, mongoDisconnect } from '~/utils/mongo.server'
+  mongoConnect,
+  mongoCountUser,
+  mongoDisconnect,
+  mongoGetProfiles,
+  mongoGetUser,
+  mongoSaveUser,
+  mongoUpdateUserLogin
+} from '~/utils/mongo.server'
 
 export async function register(email, password) {
   const emailHash = crypto.createHash('sha256').update(email).digest('hex')
   const passwordHash = await bcrypt.hash(password, 10)
   const client = await mongoConnect()
   try {
-    const res = await saveUser(client, emailHash, passwordHash)
-    if (!res.success !== true) return null
+    const user = await mongoCountUser(client, emailHash)
+    if (user !== 0) {
+      return {
+        success: false,
+        error: 'User existed'
+      }
+    }
+    const data = {
+      email_hash: emailHash,
+      last_login: Date.now(),
+      password: passwordHash,
+      profiles: []
+    }
+    await mongoSaveUser(client, data)
     return { userEmail: email }
+  } catch (err) {
+    throw new Response('register failed:' + JSON.stringify(err), {
+      status: 500
+    })
   } finally {
     await mongoDisconnect(client)
   }
@@ -27,14 +44,18 @@ export async function login(email, password) {
   const emailHash = crypto.createHash('sha256').update(email).digest('hex')
   const client = await mongoConnect()
   try {
-    const user = await getUser(client, emailHash)
+    const user = await mongoGetUser(client, emailHash)
     if (user.password === undefined) return null
     const isCorrectPassword = await bcrypt.compare(password, user.password)
     if (!isCorrectPassword) return null
     // save login time
-    const res = await updateUserLogin(client, emailHash)
+    const res = await mongoUpdateUserLogin(client, emailHash)
     if (res.success !== true) return null
     return { userEmail: email }
+  } catch (err) {
+    throw new Response('login failed:' + JSON.stringify(err), {
+      status: 500
+    })
   } finally {
     await mongoDisconnect(client)
   }
@@ -90,7 +111,15 @@ export async function retrieveUser(request) {
   const emailHash = crypto.createHash('sha256').update(userEmail).digest('hex')
   const client = await mongoConnect()
   try {
-    return await getUserWithProfile(client, emailHash)
+    const user = await mongoGetUser(client, emailHash)
+    if (user?.profiles.length !== 0) {
+      user.profiles = await mongoGetProfiles(client, user.profiles)
+    }
+    return user
+  } catch (err) {
+    throw new Response('retrieveUser failed:' + JSON.stringify(err), {
+      status: 500
+    })
   } finally {
     await mongoDisconnect(client)
   }
@@ -100,8 +129,12 @@ export async function checkUser(email) {
   const emailHash = crypto.createHash('sha256').update(email).digest('hex')
   const client = await mongoConnect()
   try {
-    const res = await countUser(emailHash)
+    const res = await mongoCountUser(client, emailHash)
     return res !== 0
+  } catch (err) {
+    throw new Response('checkUser failed:' + JSON.stringify(err), {
+      status: 500
+    })
   } finally {
     await mongoDisconnect(client)
   }
